passcrypt: Use long rather than int for random() return value
This matches the random(3) definition even though with the % 64, we only
get a number in the range 0-63 back.

+++ tools/passcrypt.c
@@ -39,7 +39,7 @@ int main(int argc, char **argv)
 	}
 
 	for (i = 3; i < 19; i++) {
-		int r;
+		long r;
 		struct timeval tv;
 
 		gettimeofday(&tv, NULL);
